+++ source/Plugins/LanguageRuntime/ObjC/CMakeLists.txt
@@ -1,1 +1,10 @@
+add_lldb_library(lldbPluginObjCRuntime PLUGIN
+  ObjCLanguageRuntime.cpp
+
+  LINK_LIBS
+    lldbCore
+    lldbSymbol
+    lldbTarget
+    lldbUtility
+)
 add_subdirectory(AppleObjCRuntime)
